refactor pipeline net
