+++ pyzoo/zoo/pipeline/api/net/graph_net.py
@@ -1,0 +1,93 @@
+#
+# Copyright 2018 Analytics Zoo Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import sys
+
+from bigdl.nn.layer import Model as BModel
+from bigdl.util.common import to_list, callBigDlFunc
+from zoo.pipeline.api.keras.engine.topology import ZooKerasLayer
+from bigdl.nn.layer import Layer
+
+if sys.version >= '3':
+    long = int
+    unicode = str
+
+
+class GraphNet(BModel):
+    def __init__(self, input, output, jvalue=None, bigdl_type="float", **kwargs):
+        super(BModel, self).__init__(jvalue,
+                                     to_list(input),
+                                     to_list(output),
+                                     bigdl_type,
+                                     **kwargs)
+
+    def flattened_layers(self, include_container=False):
+        jlayers = callBigDlFunc(self.bigdl_type, "getFlattenSubModules", self, include_container)
+        layers = [Layer.of(jlayer) for jlayer in jlayers]
+        return layers
+
+    @property
+    def layers(self):
+        jlayers = callBigDlFunc(self.bigdl_type, "getSubModules", self)
+        layers = [Layer.of(jlayer) for jlayer in jlayers]
+        return layers
+
+    @staticmethod
+    def from_jvalue(jvalue, bigdl_type="float"):
+        """
+        Create a Python Model base on the given java value
+
+        :param jvalue: Java object create by Py4j
+        :return: A Python Model
+        """
+        model = GraphNet([], [], jvalue=jvalue, bigdl_type=bigdl_type)
+        model.value = jvalue
+        return model
+
+    def new_graph(self, outputs):
+        """
+        Specify a list of nodes as output and return a new graph using the existing nodes
+
+        :param outputs: A list of nodes specified
+        :return: A graph model
+        """
+        value = callBigDlFunc(self.bigdl_type, "newGraph", self.value, outputs)
+        return self.from_jvalue(value, self.bigdl_type)
+
+    def freeze_up_to(self, names):
+        """
+        Freeze the model from the bottom up to the layers specified by names (inclusive).
+        This is useful for finetuning a model
+
+        :param names: A list of module names to be Freezed
+        :return: current graph model
+        """
+        callBigDlFunc(self.bigdl_type, "freezeUpTo", self.value, names)
+
+    def unfreeze(self, names=None):
+        """
+        "unfreeze" module, i.e. make the module parameters(weight/bias, if exists)
+        to be trained(updated) in training process.
+        If 'names' is a non-empty list, unfreeze layers that match given names
+
+        :param names: list of module names to be unFreezed. Default is None.
+        :return: current graph model
+        """
+        callBigDlFunc(self.bigdl_type, "unFreeze", self.value, names)
+
+    def to_keras(self):
+        value = callBigDlFunc(self.bigdl_type, "netToKeras", self.value)
+        return ZooKerasLayer.of(value, self.bigdl_type)
